+++ packages/iov-crypto/src/index.ts
@@ -19,4 +19,11 @@ export { Ripemd160 } from "./ripemd";
 export { Secp256k1, Secp256k1Keypair } from "./secp256k1";
 export { ExtendedSecp256k1Signature, Secp256k1Signature } from "./secp256k1signature";
 export { Sha1, Sha256, Sha512 } from "./sha";
-export { Slip10, Slip10Curve, Slip10RawIndex, Slip10Result, slip10CurveFromString } from "./slip10";
+export {
+  pathToString,
+  Slip10,
+  Slip10Curve,
+  Slip10RawIndex,
+  Slip10Result,
+  slip10CurveFromString,
+} from "./slip10";