+++ packages/iov-crypto/src/slip10.ts
@@ -183,3 +183,12 @@ export class Slip10 {
     }
   }
 }
+
+export function pathToString(path: readonly Slip10RawIndex[]): string {
+  return path.reduce((current, component): string => {
+    const componentString = component.isHardened()
+      ? `${component.toNumber() - 2 ** 31}'`
+      : component.toString();
+    return current + "/" + componentString;
+  }, "m");
+}