+++ packages/iov-crypto/types/slip10.d.ts
@@ -36,3 +36,4 @@ export declare class Slip10 {
   private static isGteN;
   private static n;
 }
+export declare function pathToString(path: readonly Slip10RawIndex[]): string;